server/requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
fastapi
fastapi[standard]
uvicorn
bcrypt
python-jose[cryptography]
passlib[bcrypt]
cryptography